Forward requests only when they have the correct buffer version
If a request with version X arrives before the textDocument/didChange
notification for version X, then the server operates on stale data.
So far we've worked around this by sending textDocument/didChange
before every relevant request, and blocking until that's done.
This is ugly and the blocking seems to be causing problems (see the
grandchild commit).

Let's forward requests only once their buffer version matches the
one we have sent to the server.

Also, drop stale requests. Not sure if this is 100% safe, we'll see.